requirements.txt
selenium==4.17.2
python-dotenv==1.0.1